requirements.txt
discord.py[voice]==2.3.2
PyNaCl==1.5.0
yt-dlp>=2024.8.6
python-dotenv==1.0.0
aiohttp==3.9.4
async-timeout==4.0.3
browser-cookie3>=0.19.1
